test(encoding): fix failing tests due to the yaml library update
